FIX : 다른 라우트에서 모달을 여는 액션을 취했을 때 열리지 않는 버그 수정
